services/ai/AiAgent/requirements.txt
# API client
openai>=1.12.0
httpx>=0.27.0

# Image processing
Pillow>=10.0.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Web server for API
flask>=2.0.0
flask-cors>=3.0.0
